add delete Owner

+++ back-end/src/main/java/sit/int204/actionback/service/EventCategoryOwnerService.java
@@ -2,17 +2,23 @@ package sit.int204.actionback.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import org.springframework.web.server.ResponseStatusException;
+import sit.int204.actionback.entities.EventCategoryOwner;
 import sit.int204.actionback.repo.EventCategoryOwnerRepository;
-import sit.int204.actionback.repo.EventCategoryRepository;
+
+import java.util.Optional;
+
 @Service
 public class EventCategoryOwnerService {
     @Autowired
     public EventCategoryOwnerRepository eventCategoryOwnerRepository;
 
     public boolean deleteForOwner(Integer user_id){
-        eventCategoryOwnerRepository.deleteOwner(user_id);
+        Optional<EventCategoryOwner> owner = eventCategoryOwnerRepository.findById(user_id);
+        if (!owner.isEmpty()) {
+            eventCategoryOwnerRepository.deleteOwner(user_id);
+        }
         return true;
     }
 }
